'cvs export' doesn't seem to work properly under Windows NT. Use 'cvs co' instead and then remove CVS directories.
git-svn-id: file:///c/gobo/svn/gobo/trunk@1344 fc6a2fa8-d027-0410-966b-9b90f175b7b2

+++ work/script/make_delivery.sh
@@ -6,17 +6,15 @@
 rootdir=`pwd`
 
 # Checkout source code from CVS repository.
-cvs export -D now gobo
+cvs co -P gobo
 
 # Set environment variable $GOBO.
-cd gobo
-GOBO=`pwd`
+GOBO=$rootdir/gobo
 GOBO=`echo $GOBO | sed "s/^\/\/\([a-zA-Z]\)\//\1:\//g"`
 export GOBO
 echo ''; echo "#### \$GOBO = $GOBO"
 
 # Get binaries to allow bootstraping.
-cd $rootdir
 cd $2
 cp * $GOBO/bin
 cd $rootdir
@@ -26,9 +24,12 @@ $GOBO/work/script/make_gobo.sh $1
 
 # Remove develop files.
 echo ''; echo "### Removing develop files..."
-cd $GOBO
-rm -rf work
-rm -rf doc/structure/visio
-rm -rf doc/time/visio
-find . -name ".cvsignore" -print -exec rm -f {} \;
+if [ -d $GOBO ]; then
+	cd $GOBO
+	rm -rf work
+	rm -rf doc/structure/visio
+	rm -rf doc/time/visio
+	find . -name ".cvsignore" -print -exec rm -f {} \;
+	find . -name "CVS" -exec rm -rf {} \; > /dev/null 2>&1
+fi
 cd $rootdir
